requirements.txt
potassium
torch
torchvision
pillow